refactor: Menu, MoveCommand 패키지 이동
