+++ src/test/java/domain/MoveCommandTest.java
@@ -1,10 +1,10 @@
 package domain;
 
+import controller.MoveCommand;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import view.MoveCommand;
 
 class MoveCommandTest {
     @DisplayName("출발지와 목적지 입력이 올바른지 검증한다.")
