+++ src/test/java/view/MenuTest.java
@@ -1,5 +1,6 @@
 package view;
 
+import controller.Menu;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
